Exercis solution - Adding/Editing Movie details, Client-side/Server-side validation, Change Data Types of Movie/Genre Model, Fix migrations

+++ Vidly/Models/GenreType.cs
@@ -7,7 +7,7 @@ namespace Vidly.Models
 {
     public class GenreType
     {
-        public short Id { get; set; }
+        public byte Id { get; set; }
         public string Name { get; set; }
     }
 }
